Create Practical links provider
